ci: install dependencies and fix config (#8)
* ci: replace template variable with value

I copy and pasted from the template, not realizing that the `$default-branch` variable wouldn't be expanded at evaluation time but was intended for generating the config file itself.

* ci: install playwright browsers and collect reports

* ci: install SurrealDB

This looks like it'll also run it, which I don't need right now since I
do it within the Playwright tests. Maybe this would be better, though?